third_party/blink/web_tests/wpt_internal/css/css-transitions/starting-style-rule-none.html

<!DOCTYPE html>
<link rel="help" href="https://github.com/w3c/csswg-drafts/issues/8174">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/css-transitions/support/helper.js"></script>
<div id="target"></div>
<style>
  #target {
    transition-property: background-color;
    transition-duration: 100s;
    transition-timing-function: steps(2, start);
    background-color: green;
  }
  @starting-style {
    #target {
      display: none;
      background-color: red;
    }
  }
</style>
<script>
  promise_test(async t => {
    await waitForAnimationFrames(2);
    assert_equals(getComputedStyle(target).backgroundColor, "rgb(0, 128, 0)",
                  "No transition of background-color when @starting-style display is 'none'");
  }, "@starting-style with display:none");
</script>
